throttle events and reorganize helper code
