+++ src/components/containers/victory-brush-container.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { VictoryContainer, Selection } from "victory-core";
-import Helpers from "./helper-methods";
-import { assign, isEqual, isFunction } from "lodash";
+import BrushHelpers from "./brush-helpers";
+import { assign, isEqual } from "lodash";
 
 
 export default class VictoryBrushContainer extends VictoryContainer {
@@ -40,145 +40,21 @@ export default class VictoryBrushContainer extends VictoryContainer {
   static defaultEvents = [{
     target: "parent",
     eventHandlers: {
-      onMouseDown: (evt, targetProps) => { // eslint-disable-line max-statements
-        evt.preventDefault();
-        const {
-          dimension, selectedDomain, domain, handleWidth, onDomainChange, cachedSelectedDomain
-        } = targetProps;
-        const fullDomainBox = targetProps.fullDomainBox ||
-          Helpers.getDomainBox(targetProps, domain);
-        const currentDomain = isEqual(selectedDomain, cachedSelectedDomain) ?
-          targetProps.currentDomain || selectedDomain || domain : selectedDomain || domain;
-        const {x, y} = Selection.getSVGEventCoordinates(evt);
-        // Ignore events that occur outside of the maximum domain region
-        if (!Helpers.withinBounds({x, y}, fullDomainBox, handleWidth)) {
-          return {};
-        }
-        const domainBox = Helpers.getDomainBox(targetProps, domain, currentDomain);
-        const activeHandles = Helpers.getActiveHandles({x, y}, targetProps, domainBox);
-        // If the event occurs in any of the handle regions, start a resize
-        if (activeHandles) {
-          return [{
-            target: "parent",
-            mutation: () => {
-              return {
-                isSelecting: true, domainBox, fullDomainBox,
-                cachedSelectedDomain: selectedDomain, currentDomain,
-                ...Helpers.getResizeMutation(domainBox, activeHandles)
-              };
-            }
-          }];
-        } else if (
-            Helpers.withinBounds({x, y}, domainBox) &&
-            !isEqual(domain, currentDomain)
-          ) {
-          // if the event occurs within a selected region start a panning event, unless the whole
-          // domain is selected
-          return [{
-            target: "parent",
-            mutation: () => ({
-              isPanning: true, startX: x, startY: y, domainBox, fullDomainBox, currentDomain,
-              cachedSelectedDomain: selectedDomain
-            })
-          }];
-        } else {
-          // if the event occurs outside the region, or if the whole domain is selected,
-          // start a new selection
-          const minimumDomain = Helpers.getMinimumDomain();
-          if (isFunction(onDomainChange)) {
-            onDomainChange(minimumDomain);
-          }
-          return [{
-            target: "parent",
-            mutation: () => {
-              return {
-                isSelecting: true, domainBox, fullDomainBox,
-                cachedSelectedDomain: selectedDomain,
-                currentDomain: Helpers.getMinimumDomain(),
-                ...Helpers.getSelectionMutation({x, y}, domainBox, dimension)
-              };
-            }
-          }];
-        }
+      onMouseDown: (evt, targetProps) => {
+        BrushHelpers.getHandlers().onMouseMove.cancel();
+        return BrushHelpers.getHandlers().onMouseDown(evt, targetProps);
       },
-      onMouseMove: (evt, targetProps) => { // eslint-disable-line max-statements
-        // if a panning or selection has not been started, ignore the event
-        if (!targetProps.isPanning && !targetProps.isSelecting) {
-          return {};
-        }
-        const {
-          dimension, scale, isPanning, isSelecting, fullDomainBox, onDomainChange
-        } = targetProps;
-        const {x, y} = Selection.getSVGEventCoordinates(evt);
-         // Ignore events that occur outside of the maximum domain region
-        if (!Helpers.withinBounds({x, y}, fullDomainBox)) {
-          return {};
-        }
-        if (isPanning) {
-          const {startX, startY} = targetProps;
-          const pannedBox = Helpers.panBox(targetProps, {x, y});
-          const constrainedBox = Helpers.constrainBox(pannedBox, fullDomainBox);
-          const currentDomain = Selection.getBounds({...constrainedBox, scale});
-          if (isFunction(onDomainChange)) {
-            onDomainChange(currentDomain);
-          }
-          return [{
-            target: "parent",
-            mutation: () => {
-              return {
-                currentDomain,
-                startX: pannedBox.x2 >= fullDomainBox.x2 || pannedBox.x1 <= fullDomainBox.x1 ?
-                  startX : x,
-                startY: pannedBox.y2 >= fullDomainBox.y2 || pannedBox.y1 <= fullDomainBox.y1 ?
-                  startY : y,
-                ...constrainedBox
-              };
-            }
-          }];
-        } else if (isSelecting) {
-          const x2 = dimension !== "y" ? x : targetProps.x2;
-          const y2 = dimension !== "x" ? y : targetProps.y2;
-          const currentDomain =
-            Selection.getBounds({x2, y2, x1: targetProps.x1, y1: targetProps.y1, scale});
-          if (isFunction(onDomainChange)) {
-            onDomainChange(currentDomain);
-          }
-          return [{
-            target: "parent",
-            mutation: () => {
-              return {
-                x2, y2, currentDomain
-              };
-            }
-          }];
-        }
+      onMouseMove: (evt, targetProps) => {
+        evt.persist();
+        return BrushHelpers.getHandlers().onMouseMove(evt, targetProps);
       },
       onMouseUp: (evt, targetProps) => {
-        const {x1, y1, x2, y2, domain, onDomainChange} = targetProps;
-        // if the mouse hasn't moved since a mouseDown event, select the whole domain region
-        if (x1 === x2 || y1 === y2) {
-          if (isFunction(onDomainChange)) {
-            onDomainChange(domain);
-          }
-          return [{
-            target: "parent",
-            mutation: () => {
-              return {
-                isPanning: false, isSelecting: false, currentDomain: domain
-              };
-            }
-          }];
-        }
-        return [{
-          target: "parent",
-          mutation: () => ({ isPanning: false, isSelecting: false })
-        }];
+        BrushHelpers.getHandlers().onMouseMove.cancel();
+        return BrushHelpers.getHandlers().onMouseUp(evt, targetProps);
       },
-      onMouseLeave: () => {
-        return [{
-          target: "parent",
-          mutation: () => ({ isPanning: false, isSelecting: false })
-        }];
+      onMouseLeave: (evt, targetProps) => {
+        BrushHelpers.getHandlers().onMouseMove.cancel();
+        return BrushHelpers.getHandlers().onMouseLeave(evt, targetProps);
       }
     }
   }];
