+++ src/components/containers/victory-selection-container.js
@@ -1,6 +1,6 @@
 import React from "react";
-import { VictoryContainer, Selection } from "victory-core";
-import { isFunction, assign } from "lodash";
+import { VictoryContainer } from "victory-core";
+import SelectionHelpers from "./selection-helpers";
 
 
 export default class VictorySelectionContainer extends VictoryContainer {
@@ -29,72 +29,16 @@ export default class VictorySelectionContainer extends VictoryContainer {
     target: "parent",
     eventHandlers: {
       onMouseDown: (evt, targetProps) => {
-        evt.preventDefault();
-        const { dimension, scale } = targetProps;
-        const {x, y} = Selection.getSVGEventCoordinates(evt);
-        const x1 = dimension !== "y" ? x : Selection.getDomainCoordinates(scale).x[0];
-        const y1 = dimension !== "x" ? y : Selection.getDomainCoordinates(scale).y[0];
-        const x2 = dimension !== "y" ? x : Selection.getDomainCoordinates(scale).x[1];
-        const y2 = dimension !== "x" ? y : Selection.getDomainCoordinates(scale).y[1];
-        if (isFunction(targetProps.onSelectionCleared)) {
-          targetProps.onSelectionCleared();
-        }
-        return [
-          {
-            target: "parent",
-            mutation: () => {
-              return {x1, y1, select: true, x2, y2};
-            }
-          }, {
-            target: "data",
-            childName: targetProps.children ? "all" : undefined,
-            eventKey: "all",
-            mutation: () => null
-          }
-        ];
+        SelectionHelpers.getHandlers().onMouseMove.cancel();
+        return SelectionHelpers.getHandlers().onMouseDown(evt, targetProps);
       },
       onMouseMove: (evt, targetProps) => {
-        const {dimension, scale, select} = targetProps;
-        if (!select) {
-          return {};
-        } else {
-          const {x, y} = Selection.getSVGEventCoordinates(evt);
-          const x2 = dimension !== "y" ? x : Selection.getDomainCoordinates(scale).x[1];
-          const y2 = dimension !== "x" ? y : Selection.getDomainCoordinates(scale).y[1];
-          return {
-            target: "parent",
-            mutation: () => {
-              return { x2, y2 };
-            }
-          };
-        }
+        evt.persist();
+        return SelectionHelpers.getHandlers().onMouseMove(evt, targetProps);
       },
       onMouseUp: (evt, targetProps) => {
-        const {x2, y2} = targetProps;
-        const parentMutation = [{
-          target: "parent",
-          mutation: () => {
-            return { select: false, x1: null, x2: null, y1: null, y2: null };
-          }
-        }];
-        if (!x2 || !y2) {
-          return parentMutation;
-        }
-        const bounds = Selection.getBounds(targetProps);
-        const datasets = Selection.getDatasets(targetProps);
-        const selectedData = Selection.filterDatasets(datasets, bounds);
-        const callbackMutation = selectedData && isFunction(targetProps.onSelection) ?
-          targetProps.onSelection(selectedData, bounds) : {};
-        const dataMutation = selectedData ?
-          selectedData.map((d) => {
-            return {
-              childName: d.childName, eventKey: d.eventKey, target: "data",
-              mutation: () => {
-                return assign({active: true}, callbackMutation);
-              }
-            };
-          }) : [];
-        return parentMutation.concat(dataMutation);
+        SelectionHelpers.getHandlers().onMouseMove.cancel();
+        return SelectionHelpers.getHandlers().onMouseUp(evt, targetProps);
       }
     }
   }];
